test(rig): keep command path coverage out of pipeline core
