+++ src/core/rig/pipeline.rs
@@ -575,99 +575,6 @@ fn create_symlink(_target: &Path, _link: &Path) -> std::io::Result<()> {
     ))
 }
 
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::fs;
-
-    use super::run_pipeline;
-    use crate::rig::spec::{PipelineStep, RigSpec};
-    use crate::rig::toolchain;
-
-    fn rig_with_command(cmd: String, env: HashMap<String, String>) -> RigSpec {
-        let mut pipeline = HashMap::new();
-        pipeline.insert(
-            "up".to_string(),
-            vec![PipelineStep::Command {
-                step_id: None,
-                depends_on: Vec::new(),
-                cmd,
-                cwd: None,
-                env,
-                label: Some("command".to_string()),
-            }],
-        );
-
-        RigSpec {
-            id: "command-env-test".to_string(),
-            description: String::new(),
-            components: Default::default(),
-            services: Default::default(),
-            symlinks: Vec::new(),
-            shared_paths: Vec::new(),
-            resources: Default::default(),
-            pipeline,
-            bench: None,
-            app_launcher: None,
-            bench_workloads: Default::default(),
-        }
-    }
-
-    #[test]
-    fn test_command_step_explicit_path_env_wins() {
-        let tmp = tempfile::tempdir().expect("tmpdir");
-        let path_file = tmp.path().join("path.txt");
-        let mut env = HashMap::new();
-        env.insert("PATH".to_string(), "/tmp/explicit-toolchain".to_string());
-
-        let rig = rig_with_command(
-            format!("printf '%s' \"$PATH\" > {}", path_file.to_string_lossy()),
-            env,
-        );
-        let outcome = run_pipeline(&rig, "up", true).expect("pipeline");
-
-        assert!(outcome.is_success(), "outcomes: {:?}", outcome.steps);
-        assert_eq!(
-            fs::read_to_string(path_file).expect("path file"),
-            "/tmp/explicit-toolchain"
-        );
-    }
-
-    #[test]
-    fn test_command_step_uses_bootstrapped_toolchain_path() {
-        let expected = toolchain::command_step_path().expect("toolchain path");
-        let tmp = tempfile::tempdir().expect("tmpdir");
-        let path_file = tmp.path().join("path.txt");
-
-        let rig = rig_with_command(
-            format!("printf '%s' \"$PATH\" > {}", path_file.to_string_lossy()),
-            HashMap::new(),
-        );
-        let outcome = run_pipeline(&rig, "up", true).expect("pipeline");
-
-        assert!(outcome.is_success(), "outcomes: {:?}", outcome.steps);
-        assert_eq!(
-            fs::read_to_string(path_file).expect("path file"),
-            expected.to_string_lossy()
-        );
-    }
-
-    #[test]
-    fn test_command_step_exit_127_mentions_path_contract() {
-        let rig = rig_with_command(
-            "definitely-not-a-homeboy-test-command-1758 2>/dev/null".to_string(),
-            HashMap::new(),
-        );
-        let outcome = run_pipeline(&rig, "up", true).expect("pipeline report");
-
-        assert!(!outcome.is_success());
-        let error = outcome.steps[0].error.as_deref().expect("error");
-        assert!(error.contains("exited 127"), "{error}");
-        assert!(error.contains("command not found"), "{error}");
-        assert!(error.contains("env.PATH"), "{error}");
-    }
-}
-
 fn verify_symlink(rig: &RigSpec, link: &SymlinkSpec) -> Result<()> {
     let link_path = PathBuf::from(expand_vars(rig, &link.link));
     let target_path = PathBuf::from(expand_vars(rig, &link.target));
